chatbot/requirements.txt
langgraph>=0.0.1
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langgraph-checkpoint-sqlite>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
